refactor: :recycle: First commit after starting to refactor CSS code and start using React Context
1- Create first SASS files, first global styles with a reset of Modern Normalize. 2- Create first context files successful for Home mainly for the Modal that is responsible for configuring new timers
