+++ src/components/getTimers/GetTimers.js
@@ -1,0 +1,9 @@
+/* import { useLocalStorage } from "../../utils/hooks/useLocalStorage";
+
+const GetTimers = (itemName, initialValue) => {
+  const [timers, setTimers] = useLocalStorage(itemName, initialValue);
+
+  // return [timers, setTimers];
+};
+
+export default GetTimers; */
