+++ src/context/newTimerModal/NewTimerModalContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const NewTimerModalContext = createContext();
+
+export default NewTimerModalContext;
