+++ src/context/newTimerModal/NewTimerModalReducer.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const NewTimerModalReducer = () => {
+  return <div>NewTimerModalReducer</div>;
+};
+
+export default NewTimerModalReducer;
