+++ src/context/timer/TimerContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const TimerContext = createContext();
+
+export default TimerContext;
