+++ src/context/timer/TimerReducer.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const TimerReducer = () => {
+  return (
+    <div>TimerReducer</div>
+  )
+}
+
+export default TimerReducer
