+++ src/styles/main.module.scss
@@ -1,0 +1,3 @@
+@import "./sass/globals";
+
+// CAMBIAR LA MAYORIA DE LAS UNIDADES DE MEDIDADA A UN 1EM QUE SIGNIFICA EL TAMAÑO EN FONTZISE DEL CONTENEDOR MAS CERCANO O DE EL MISMO Y UN 1REM QUE SIGNIFICA EL FONTZISE DEL GLOBAL QUE ES UN CALCULO INTERESANTE PERO QUE GENERALMENTE ES DE 16PX PERO EN REACT ESTA MEJOR CONFIGURADO, EN REACT MEJOR MANEJAR UN REM EN GENERAL O MEJOR IR REVISANDO LAS NECESIDADES REPERZONALIZANDO LOS ESTILOS PARA VER COMO QUEDA TODO MEJOR DE FORMA RESPONSIVE  Y ACORDATE DE SCALE PARA REDUCIR TAMAÑOS ESPECIFICOS, PROBALO A VER SI FUNCIONA CON LOS REMS, A Y TAMBIEN EL GAP PARA HACER UN ESPACIADO COMO MARGIN CREO IGUAL PROBALO, Y TAMBIEN WRAPPER PARA FLEXBOX PARA QUE PERMITA MODIFICAR EL WIDTH, Y BASICO QUE DISPLAY FLEX HACE QUE EL CONTENEDOR SOLO OCUPE EL ESPACIO DE SU CONTENIDO
