+++ src/styles/sass/_globals.scss
@@ -1,3 +1,5 @@
+@import "./modernNormalize";
+
 .App {
   min-height: 100vh;
   text-align: center;
