+++ src/views/timersDetails/TestContextInTimerDetails.js
@@ -1,0 +1,20 @@
+import { useEffect, useContext } from "react";
+import TimerContext from "../../context/timer/TimerContext";
+
+// CONTEXT
+// import TimerState from "../../context/timer/TimerState";
+// import TestContextInTimerDetails from "./TestContextInTimerDetails";
+
+const TestContextInTimerDetails = () => {
+  const { getTimers, getTimer } =
+    useContext(TimerContext);
+
+  useEffect(() => {
+    getTimers();
+    getTimer(1);
+  }, []);
+
+  // return <TimerState></TimerState>;
+};
+
+export default TestContextInTimerDetails;
