+++ src/views/timersDetails/TimersDetails.js
@@ -6,6 +6,7 @@ import Button from "react-bootstrap/Button";
 import ResetButton from "../../components/resetButton/ResetButton";
 import LockButton from "../../components/lockButton/LockButton";
 
+
 const TimersDetails = () => {
   return (
     <div>
